Unseal Node, move inner classes to top level
