+++ src/r3/atomic-swap/corda/evm-interop-common/src/main/kotlin/com/r3/corda/interop/evm/common/trie/NibbleArray.kt
@@ -1,3 +1,19 @@
+/*
+ * Copyright 2023, R3 LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package com.r3.corda.interop.evm.common.trie
 
 /**
@@ -62,8 +78,6 @@ data class NibbleArray(val values: ByteArray) {
     val tail: NibbleArray get() = dropFirst(1)
 
     fun isEmpty(): Boolean = values.isEmpty()
-
-    fun isNotEmpty(): Boolean = values.isNotEmpty()
 
     fun startsWith(other: NibbleArray): Boolean {
         if (other.size > size) {
