+++ src/r3/atomic-swap/corda/evm-interop-common/src/main/kotlin/com/r3/corda/interop/evm/common/trie/Node.kt
@@ -18,35 +18,33 @@ package com.r3.corda.interop.evm.common.trie
 
 import org.web3j.crypto.Hash.sha3
 import org.web3j.rlp.*
-import org.web3j.utils.Numeric
 
 /**
  * The base class for all types of nodes in a Patricia Trie.
  */
-sealed class Node {
+interface Node {
     /**
      * Get the encoded version of this node.
      * @return encoded byte array of this node.
      */
-    abstract val encoded: ByteArray
+    val encoded: ByteArray
 
     /**
      * Get the SHA-3 hash of the encoded node.
      * @return hash of the encoded node.
      */
-    open val hash: ByteArray
+    val hash: ByteArray
         get() = sha3(encoded)
 
     /**
      * Puts a key-value pair into a node of the Patricia Trie.
      * If the node does not exist, a new LeafNode is created.
      *
-     * @param node The node where to put the key-value pair.
      * @param key The key to put.
-     * @param value The value to put.
+     * @param newValue The value to put.
      * @return The node where the key-value pair was put.
      */
-    abstract fun put(key: NibbleArray, newValue: ByteArray): Node
+    fun put(key: NibbleArray, newValue: ByteArray): Node
 
     /**
      * Gets the value for a given key from the Patricia Trie.
@@ -54,7 +52,7 @@ sealed class Node {
      * @param key The key for which to get the value.
      * @return The value associated with the key, or an empty ByteArray if the key does not exist.
      */
-    abstract fun get(key: NibbleArray): ByteArray
+    fun get(key: NibbleArray): ByteArray
 
     /**
      * Generates a Merkle proof for a given key.
@@ -63,23 +61,13 @@ sealed class Node {
      * @param store A simple Key-Value that will collect the trie proofs
      * @return Merkle proof as KeyValueStore.
      */
-    abstract fun generateMerkleProof(key: NibbleArray, store: WriteableKeyValueStore) : KeyValueStore
+    fun generateMerkleProof(key: NibbleArray, store: WriteableKeyValueStore) : KeyValueStore
 
-    abstract fun verifyMerkleProof(
+    fun verifyMerkleProof(
         key: NibbleArray,
         expectedValue: ByteArray,
         proof: KeyValueStore
     ): Boolean
-
-    /**
-     * Provide a String representation of the Node for debugging.
-     * @return String representation of the Node.
-     */
-    override fun toString(): String {
-        return "NodeType: ${javaClass.simpleName} " +
-                "Hash: ${Numeric.toHexString(hash)} " +
-                "Encoded: ${Numeric.toHexString(encoded)}"
-    }
 
     companion object {
 
@@ -88,9 +76,6 @@ sealed class Node {
 
         fun leaf(key: ByteArray, value: ByteArray): LeafNode = leaf(NibbleArray.fromBytes(key), value)
         fun leaf(key: NibbleArray, value: ByteArray): LeafNode = LeafNode(key, value)
-
-        fun branch(): BranchNode = branch(emptyArray)
-        fun branch(value: ByteArray) = branch(Array<Node>(16) { EmptyNode }, value)
 
         fun branch(sparseBranches: List<Pair<Int, Node>>): BranchNode = branch(sparseBranches, emptyArray)
         fun branch(sparseBranches: List<Pair<Int, Node>>, value: ByteArray): BranchNode {
@@ -101,7 +86,7 @@ sealed class Node {
             return branch(branches, value)
         }
 
-        private fun branch(branches: Array<Node>, value: ByteArray): BranchNode = BranchNode(branches, value)
+        fun branch(branches: Array<Node>, value: ByteArray): BranchNode = BranchNode(branches, value)
 
         fun extension(path: NibbleArray, value: Node): ExtensionNode = ExtensionNode(path, value)
 
@@ -177,309 +162,4 @@ sealed class Node {
         }
     }
 
-    /**
-     * Represents an EmptyNode in a Patricia Trie.
-     *
-     * This class is a specific type of Node, with its encoded form
-     * being the RLP (Recursive Length Prefix) encoding of an empty byte array.
-     */
-    object EmptyNode : Node() {
-        /**
-         * Returns the RLP-encoded form of the EmptyNode,
-         * which is an empty byte array encoded in RLP.
-         */
-        override val encoded: ByteArray
-            get() = RlpEncoder.encode(RlpString.create(ByteArray(0)))
-
-        override fun put(key: NibbleArray, newValue: ByteArray): Node = leaf(key, newValue)
-
-        override fun get(key: NibbleArray): ByteArray = ByteArray(0)
-
-        override fun generateMerkleProof(key: NibbleArray, store: WriteableKeyValueStore): KeyValueStore {
-            throw IllegalArgumentException("Key is not part of the trie")
-        }
-
-        override fun verifyMerkleProof(key: NibbleArray, expectedValue: ByteArray, proof: KeyValueStore): Boolean {
-            throw IllegalArgumentException("Key is not part of the trie")
-        }
-    }
-
-    /**
-     * Represents an ExtensionNode in a Patricia Trie.
-     *
-     * An ExtensionNode consists of a path and an inner node.
-     * The encoded form is an RLP (Recursive Length Prefix) encoded list of the path and the inner node.
-     *
-     * @property path The path of the ExtensionNode, stored as a nibbles array.
-     * @property innerNode The inner Node that the ExtensionNode points to.
-     */
-    class ExtensionNode(
-        val path: NibbleArray,
-        val innerNode: Node
-    ) : Node() {
-
-        /**
-         * The RLP-encoded form of the ExtensionNode, which is an RLP-encoded list of the path and the inner node.
-         */
-        override val encoded: ByteArray
-            get() {
-                val encodedInnerNode = innerNode.encoded
-                return RlpEncoder.encode(
-                    RlpList(
-                        RlpString.create(PatriciaTriePathType.EXTENSION.applyPrefix(path).toBytes()),
-                        if (encodedInnerNode.size >= 32) {
-                            RlpString.create(innerNode.hash)
-                        } else {
-                            RlpDecoder.decode(encodedInnerNode) // TODO: review
-                        }
-                    )
-                )
-            }
-
-        override fun put(key: NibbleArray, newValue: ByteArray): Node {
-            val matchingLength = path.prefixMatchingLength(key)
-
-            // Key (1, 2, 3...) contains entire path (1, 2, 3)
-            if (matchingLength == path.size) {
-                // Put the value into the inner node, at the remaining key
-                return extension(path, innerNode.put(key.dropFirst(matchingLength), newValue))
-            }
-
-            // Key (1, 2, 3...) contains part of path (1, 2, 4, 5, 6)
-            val matchingPath = path.takeFirst(matchingLength)       // Nibbles where key and path agree (1, 2)
-            val pathIndex = path[matchingLength].toInt()            // Nibble where key and path diverge (4)
-            val remainingPath = path.remainingAfter(matchingLength) // Path nibbles after divergence (5, 6)
-
-            // Branch either to the inner node or to an extension terminating in the inner node
-            val firstBranch = if (remainingPath.isEmpty()) innerNode else extension(remainingPath, innerNode)
-
-            val branchNode = if (matchingLength == key.size) {
-                // Path (1, 2, 3...) contains entire key (1, 2, 3)
-
-                // 3 -> firstBranch
-                branch(listOf(pathIndex to firstBranch), newValue)
-            } else {
-                // Path (1, 2, 3...) contains part of key (1, 2, 4, 5, 6)
-                val keyIndex = key[matchingLength].toInt()            // Nibble where key and path diverge (4)
-                val remainingKey = key.remainingAfter(matchingLength) // Key nibbles after divergence (5, 6)
-
-                // 3 -> firstBranch
-                // 4 -> leaf((5, 6), newValue)
-                branch(listOf(pathIndex to firstBranch, keyIndex to leaf(remainingKey, newValue)))
-            }
-
-            return if (matchingPath.isEmpty()) branchNode else extension(matchingPath, branchNode)
-        }
-
-        override fun get(key: NibbleArray): ByteArray {
-            val matchingLength = path.prefixMatchingLength(key)
-            if (matchingLength < path.size) {
-                return ByteArray(0) // TODO: key not found
-            }
-
-            return innerNode.get(key.dropFirst(matchingLength))
-        }
-
-        override fun generateMerkleProof(key: NibbleArray, store: WriteableKeyValueStore): KeyValueStore =
-                if (key.startsWith(path)) {
-                    store.put(hash, encoded)
-                    innerNode.generateMerkleProof(key.dropFirst(path.size), store)
-                } else {
-                    throw IllegalArgumentException("Key is not part of the trie")
-                }
-
-        override fun verifyMerkleProof(key: NibbleArray, expectedValue: ByteArray, proof: KeyValueStore): Boolean =
-            if (key.startsWith(path)) {
-                verifyMerkleProof(
-                    innerNode.hash,
-                    key.dropFirst(path.size),
-                    expectedValue,
-                    proof
-                )
-            } else {
-                throw IllegalArgumentException("Key is not part of the trie")
-            }
-
-    }
-
-    /**
-     * Represents a BranchNode in a Patricia Trie.
-     *
-     * A BranchNode consists of an array of Node branches and a value.
-     * The encoded form is an RLP (Recursive Length Prefix) encoded list of the branches and value.
-     *
-     * @property branches The array of Nodes that the BranchNode contains.
-     * @property value The value of the BranchNode.
-     */
-    class BranchNode(
-        private val branches: Array<Node>,
-        val value: ByteArray
-    ) : Node() {
-
-        /**
-         * The RLP-encoded form of the BranchNode, which is an RLP-encoded list of the branches and value.
-         */
-        override val encoded: ByteArray
-            get() {
-                return RlpEncoder.encode(RlpList(branches.map { node ->
-                    val encodedNode = node.encoded
-                    when {
-                        node is EmptyNode -> RlpString.create(ByteArray(0))
-                        // NOTE: in the next line, if node can be a HashNode then we need to call
-                        // node.hash otherwise we can optimize and call Hash.sha3(encodedNode)
-                        encodedNode.size >= 32 -> RlpString.create(node.hash)
-                        else -> RlpString.create(encodedNode)
-                    }
-                }.plus(RlpString.create(value))))
-            }
-
-        fun getBranch(branch: Byte): Node = branches[branch.toInt()]
-
-        override fun put(key: NibbleArray, newValue: ByteArray): Node =
-            if (key.isEmpty()) branch(branches, newValue)
-            else {
-                val newBranches = Array(16) { index ->
-                    if (index == key.head.toInt()) getBranch(key.head).put(key.tail, newValue) else branches[index]
-                }
-                branch(newBranches, value)
-            }
-
-        override fun get(key: NibbleArray): ByteArray = if (key.isEmpty()) value else getBranch(key.head).get(key.tail)
-
-        override fun generateMerkleProof(key: NibbleArray, store: WriteableKeyValueStore): KeyValueStore {
-                store.put(hash, encoded)
-
-                if (key.isEmpty()) {
-                    require(value.isNotEmpty()) { "Terminal branch without value" }
-                    return store
-                }
-
-                return getBranch(key.head).generateMerkleProof(key.tail, store)
-            }
-
-        override fun verifyMerkleProof(key: NibbleArray, expectedValue: ByteArray, proof: KeyValueStore): Boolean =
-            if (key.isEmpty()) {
-                value.contentEquals(expectedValue)
-            } else {
-                verifyMerkleProof(getBranch(key.head).hash, key.tail, expectedValue, proof)
-            }
-    }
-
-    /**
-     * A representation of a node in a Patricia Trie that is used to store a hash and an inner node.
-     * This class provides methods to create a new HashNode and retrieve its encoded representation.
-     *
-     * @param hash The hash of the node
-     * @param innerNode The inner node associated with this HashNode. Defaults to an EmptyNode if not provided.
-     */
-    class HashNode(override val hash: ByteArray, private val innerNode: Node) : Node() {
-        /**
-         * Returns the encoded representation of the HashNode. If the inner node is an EmptyNode,
-         * it returns the hash of the node, otherwise it returns the encoded representation of the inner node.
-         */
-        override val encoded: ByteArray get() = if (innerNode is EmptyNode) hash else innerNode.encoded
-
-        override fun put(key: NibbleArray, newValue: ByteArray): Node {
-            throw UnsupportedOperationException("Cannot put into HashNode")
-        }
-
-        override fun get(key: NibbleArray): ByteArray {
-            throw UnsupportedOperationException("Cannot get from HashNode")
-        }
-
-        override fun generateMerkleProof(key: NibbleArray, store: WriteableKeyValueStore): KeyValueStore {
-            throw UnsupportedOperationException("Cannot generate Merkle proof from HashNode")
-        }
-
-        override fun verifyMerkleProof(key: NibbleArray, expectedValue: ByteArray, proof: KeyValueStore): Boolean {
-            throw UnsupportedOperationException("Cannot verify Merkle proof from HashNode")
-        }
-    }
-
-    /**
-     * Represents a LeafNode in a Patricia Trie.
-     *
-     * A LeafNode consists of a path (represented as a nibble array) and a value (represented as a byte array).
-     * The encoded form is an RLP (Recursive Length Prefix) encoded list of the path and value.
-     *
-     * @property path The path of the node, represented as a nibble array.
-     * @property value The value of the node, represented as a byte array.
-     */
-    class LeafNode(
-        val path: NibbleArray,
-        val value: ByteArray
-    ) : Node() {
-
-        /**
-         * The RLP-encoded form of the LeafNode, which is an RLP-encoded list of the path and value.
-         */
-        override val encoded: ByteArray
-            get() {
-                return RlpEncoder.encode(
-                    RlpList(
-                        RlpString.create(PatriciaTriePathType.LEAF.applyPrefix(path).toBytes()),
-                        RlpString.create(value)
-                    )
-                )
-            }
-
-        override fun put(key: NibbleArray, newValue: ByteArray): Node {
-            // Overwrite value if key and path match exactly
-            if (path == key) return leaf(key, newValue)
-
-            val matchingLength = path.prefixMatchingLength(key)
-
-            val branches = mutableListOf<Pair<Int, Node>>()
-
-            // If there's some path (1, 2, 3, 4) left after the match (1, 2) with key (1, 2, 5, 6)
-            if (matchingLength < path.size) {
-                // Put the current value in a branch: 3 -> leaf((4), value)
-                branches.add(path[matchingLength].toInt() to leaf(path.remainingAfter(matchingLength), value))
-            }
-
-            // If there's some key (1, 2, 5, 6) left after the match (1, 2) with path (1, 2, 3, 4)
-            if (matchingLength < key.size) {
-                // Put the new value in a branch: 5 -> ((6), newValue)
-                branches.add(key[matchingLength].toInt() to leaf(key.remainingAfter(matchingLength), newValue))
-            }
-
-            /*
-            Note implicit logic here:
-
-            * matchingLength is never more than path.size, and never more than key.size
-            * path.size and key.size are never equal, or we would have hit the exact match condition above.
-            * If matchingLength is path.size, it cannot be equal to key.size, so it must be less than key.size.
-            * If matchingLength is key.size, it cannot be equal to path.size, so it must be less than path.size
-
-             */
-            val branchNode = when (matchingLength) {
-                // Implicitly, matchingLength < key.size, so branch to newValue is in branches
-                path.size -> branch(branches, value)
-                // Implicitly, matchingLength < path.size, so branch to value is in branches
-                key.size -> branch(branches, newValue)
-                // MatchingLength < key.size && matchingLength < path.size, so both branches are present
-                else -> branch(branches)
-            }
-
-            return if (matchingLength == 0) branchNode else extension(path.takeFirst(matchingLength), branchNode)
-        }
-
-        override fun get(key: NibbleArray): ByteArray = if (key == path) value else ByteArray(0)
-
-        override fun generateMerkleProof(key: NibbleArray, store: WriteableKeyValueStore): KeyValueStore {
-            if (path == key) {
-                store.put(hash, encoded)
-                return store
-            } else {
-                throw IllegalArgumentException("Key is not part of the trie")
-            }
-        }
-
-        override fun verifyMerkleProof(key: NibbleArray, expectedValue: ByteArray, proof: KeyValueStore): Boolean =
-            if (path == key) {
-                value.contentEquals(expectedValue)
-            } else {
-                throw IllegalArgumentException("Key is not part of the trie")
-            }
-    }
 }
