+++ src/r3/atomic-swap/corda/evm-interop-common/src/main/kotlin/com/r3/corda/interop/evm/common/trie/PatriciaTrie.kt
@@ -16,8 +16,6 @@
 
 package com.r3.corda.interop.evm.common.trie
 
-import com.r3.corda.interop.evm.common.trie.Node.*
-
 /**
  * The Patricia Trie is a space-optimized version of a binary trie.
  * It's an ordered tree data structure used to store a dynamic set or associative array
